Broke up MDSimulationFactory.h and various fixes
